move map tests into crate::map::tests
